+++ package.json
@@ -6,6 +6,7 @@
     "@types/express": "^4.17.13",
     "@types/jest": "^27.0.1",
     "@types/jsonwebtoken": "^8.5.5",
+    "@types/jwk-to-pem": "^2.0.1",
     "@types/node": "^16.9.0",
     "@types/pem": "^1.9.6",
     "@types/supertest": "^2.0.11",
@@ -60,7 +61,8 @@
   "dependencies": {
     "axios": "^0.21.4",
     "base-64": "^1.0.0",
-    "jsonwebtoken": "^8.5.1"
+    "jsonwebtoken": "^8.5.1",
+    "jwk-to-pem": "^2.0.5"
   },
   "publishConfig": {
     "access": "public"
