+++ src/types/JWK.ts
@@ -1,0 +1,8 @@
+export interface JWK {
+  alg: Algorithm;
+  kty: string;
+  use: string;
+  n: string;
+  e: string;
+  kid: string;
+}
